Refactor repository.
