+++ src/AppBundle/Command/LoadReportCommand.php
@@ -36,7 +36,6 @@ class LoadReportCommand extends ContainerAwareCommand
             // a good practice is to use the 'app:' prefix to group all your custom application commands
             ->setName('app:loadreport')
             ->setDescription('Loads a report');
-
     }
 
     /**
@@ -45,7 +44,6 @@ class LoadReportCommand extends ContainerAwareCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-
         $username = 'john_user';
         $file = __DIR__.'/../../../var/data/testdata.txt';
 
@@ -58,12 +56,9 @@ class LoadReportCommand extends ContainerAwareCommand
         } else {
             /** @var CampaignPerformanceProcessor $service */
             $service = $this->getContainer()->get('cpr');
-            $service->processReport($file, $account);
+            //$service->processReport($file, $account);
             //New action for reload cache table
             $service->saveCache($account->getId());
         }
-
     }
-
-
 }
