+++ src/AppBundle/Repository/AccountsRepository.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 use Doctrine\ORM\Mapping as ORM;
