+++ src/AppBundle/Repository/CampaignsRepository.php
@@ -1,8 +1,5 @@
 <?php
-
-
-
-namespace AppBundle\Entity;
+namespace AppBundle\Repository;
 
 use Doctrine\ORM\EntityRepository;
